Move extract_waveforms_to_single_buffer to tools.py
